+++ src/section/timing_points/effect_flags.rs
@@ -2,27 +2,27 @@ use std::{num::ParseIntError, str::FromStr};
 
 #[derive(Copy, Clone, Default, Debug, PartialEq, Eq)]
 /// Effect flags for a control point.
-pub struct EffectFlags(u8);
+pub struct EffectFlags(i32);
 
 impl EffectFlags {
-    pub const NONE: u8 = 0;
-    pub const KIAI: u8 = 1 << 0;
-    pub const OMIT_FIRST_BAR_LINE: u8 = 1 << 3;
+    pub const NONE: i32 = 0;
+    pub const KIAI: i32 = 1 << 0;
+    pub const OMIT_FIRST_BAR_LINE: i32 = 1 << 3;
 
     /// Check whether any of the given bitflags are set.
-    pub const fn has_flag(self, flag: u8) -> bool {
+    pub const fn has_flag(self, flag: i32) -> bool {
         (self.0 & flag) != 0
     }
 }
 
-impl From<EffectFlags> for u8 {
+impl From<EffectFlags> for i32 {
     fn from(kind: EffectFlags) -> Self {
         kind.0
     }
 }
 
-impl From<u8> for EffectFlags {
-    fn from(flags: u8) -> Self {
+impl From<i32> for EffectFlags {
+    fn from(flags: i32) -> Self {
         Self(flags)
     }
 }
